backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
python-dotenv